feat(api): folder structure and cleaner api surface for rule validations
