+++ packages/api/lib/graphql/directives/rules/index.ts
@@ -1,0 +1,7 @@
+import * as rules from './rules'
+
+import { directiveFactory } from './directiveFactory'
+import { createDirectiveTypeDefs } from './typeDefTemplate'
+
+export const ruleDirectives = directiveFactory(rules)
+export const directiveTypeDefs = createDirectiveTypeDefs(rules)
